Move tx to its own module
bacon: test
Resolves: OKTA-89669
