+++ webpack.AMDJqueryQ.config.js
@@ -28,7 +28,8 @@ module.exports = {
   plugins: [
     new webpack.DefinePlugin({
       SDK_VERSION: JSON.stringify(packageJson.version),
-      STATE_TOKEN_COOKIE_NAME: JSON.stringify(oktaAuthConfig.STATE_TOKEN_COOKIE_NAME)
+      STATE_TOKEN_COOKIE_NAME: JSON.stringify(oktaAuthConfig.STATE_TOKEN_COOKIE_NAME),
+      DEFAULT_POLLING_DELAY: oktaAuthConfig.DEFAULT_POLLING_DELAY
     })
   ]
 };
